requirements.txt
python-dotenv
requests
pandas
FastAPI
uvicorn
Optional
pytest
pydantic